Update deploy.yml
The failure in your GitHub Actions job is due to the use of a deprecated version of `actions/upload-artifact: v3`. To fix this, you need to update the action in your workflow file to use a supported version.

Replace:
```yaml
- name: Upload GitHub Pages artifact
  uses: actions/upload-pages-artifact@v2
```

With the updated action:
```yaml
- name: Upload GitHub Pages artifact
  uses: actions/upload-pages-artifact@v3
```

Here is the updated section of your workflow file:
```yaml
- name: Upload GitHub Pages artifact
  uses: actions/upload-pages-artifact@v3
  with:
    path: ${{ env.DEPLOY_DIR }} # optional, default is _site/
```

Make this change in your `.github/workflows/deploy.yml` file and re-run the workflow.